moved valid_parted_position to disk. And made it handle float numbers.
