+++ archinstall/lib/disk.py
@@ -13,6 +13,24 @@ GPT = 0b00000001
 MBR = 0b00000010
 
 
+def valid_parted_position(pos :str):
+	if not len(pos):
+		return False
+
+	if pos.isdigit():
+		return True
+
+	if pos[-1] == '%' and pos[:-1].isdigit():
+		return True
+
+	if pos[-3:].lower() in ['mib', 'kib', 'b', 'tib'] and pos[:-3].replace(".", "", 1).isdigit():
+		return True
+
+	if pos[-2:].lower() in ['kb', 'mb', 'gb', 'tb'] and pos[:-2].replace(".", "", 1).isdigit():
+		return True
+
+	return False
+
 def valid_fs_type(fstype :str) -> bool:
 	# https://www.gnu.org/software/parted/manual/html_node/mkpart.html
 	# Above link doesn't agree with `man parted` /mkpart documentation:
